+++ TrabalhoFinal/ecommerce/src/main/java/com/residencia/ecommerce/services/ClienteService.java
@@ -1,25 +1,21 @@
 package com.residencia.ecommerce.services;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
+import java.util.Optional;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
-import org.springframework.web.client.HttpClientErrorException;
-import org.springframework.web.client.HttpServerErrorException;
-import org.springframework.web.client.RestTemplate;
-import org.springframework.web.client.UnknownHttpStatusCodeException;
 
 import com.residencia.ecommerce.dto.ClienteDTO;
-import com.residencia.ecommerce.dto.ViaCepDTO;
 import com.residencia.ecommerce.entites.Cliente;
 import com.residencia.ecommerce.entites.Endereco;
-import com.residencia.ecommerce.exception.CEPNotFoundException;
 import com.residencia.ecommerce.exception.ClienteCpfDuplicadoException;
 import com.residencia.ecommerce.exception.ClienteEmailDuplicadoException;
 import com.residencia.ecommerce.exception.ClienteNotFoundException;
+import com.residencia.ecommerce.exception.CustomException;
+import com.residencia.ecommerce.exception.NoSuchElementException;
 import com.residencia.ecommerce.repositories.ClienteRepository;
 import com.residencia.ecommerce.repositories.EnderecoRepository;
 
@@ -30,7 +26,10 @@ public class ClienteService {
 
 	@Autowired
 	EnderecoRepository enderecoRepository;
-
+	
+	@Autowired
+	EnderecoService enderecoService;
+	
 	@Autowired
 	ModelMapper modelMapper;
 
@@ -43,12 +42,17 @@ public class ClienteService {
 	}
 
 	public Cliente saveCliente(Cliente cliente) {
-		Cliente clienteCpfExistente = clienteRepository.findByCpf(cliente.getCpf());
-		Cliente clienteEmailExistente = clienteRepository.findByEmail(cliente.getEmail());
-		if (clienteCpfExistente != null) {
-			throw new ClienteCpfDuplicadoException();
-		} else if (clienteEmailExistente != null) {
-			throw new ClienteEmailDuplicadoException();
+		try {
+			Cliente clienteCpfExistente = clienteRepository.findByCpf(cliente.getCpf());
+			Cliente clienteEmailExistente = clienteRepository.findByEmail(cliente.getEmail());
+			if (clienteCpfExistente != null) {
+				throw new ClienteCpfDuplicadoException();
+			} else if (clienteEmailExistente != null) {
+				throw new ClienteEmailDuplicadoException();
+			} 
+		}
+		catch(DataIntegrityViolationException ex) {
+			throw new CustomException("É necessário utilizar um endereço cadastrado posteriormente neste tipo de operação");
 		}
 		return clienteRepository.save(cliente);
 	}
@@ -58,6 +62,11 @@ public class ClienteService {
 	}
 
 	public Boolean deleteCliente(Integer id) {
+		Optional<Cliente> clienteEncontrada = clienteRepository.findById(id);
+		
+		if(clienteEncontrada.isEmpty()) {
+			throw new NoSuchElementException("Cliente com id: " + id + " não encontrada!");
+		}
 		clienteRepository.deleteById(id);
 		Cliente clienteDeletada = clienteRepository.findById(id).orElse(null);
 		return clienteDeletada == null;
@@ -74,7 +83,8 @@ public class ClienteService {
 
 	public ClienteDTO saveClienteDTO(ClienteDTO clienteDTO) {
 		Cliente cliente = modelMapper.map(clienteDTO, Cliente.class);
-
+		
+		//Verificações de duplicacao de cpf e email
 		Cliente clienteCpfExistente = clienteRepository.findByCpf(cliente.getCpf());
 		Cliente clienteEmailExistente = clienteRepository.findByEmail(cliente.getEmail());
 		if (clienteCpfExistente != null) {
@@ -82,32 +92,13 @@ public class ClienteService {
 		} else if (clienteEmailExistente != null) {
 			throw new ClienteEmailDuplicadoException();
 		}
-
-		RestTemplate restTemplate = new RestTemplate();
-		String uri = "http://viacep.com.br/ws/{cep}/json";
-		Map<String, String> params = new HashMap<>();
-		params.put("cep", clienteDTO.getCep());
-
-		try {
-			ViaCepDTO responseViaCep = restTemplate.getForObject(uri, ViaCepDTO.class, params);
-			if (responseViaCep == null || responseViaCep.getErro()) {
-				throw new CEPNotFoundException(clienteDTO.getCep());
-			}
-			Endereco endereco = new Endereco();
-			endereco.setCep(responseViaCep.getCep());
-			endereco.setRua(responseViaCep.getLogradouro());
-			endereco.setBairro(responseViaCep.getBairro());
-			endereco.setCidade(responseViaCep.getLocalidade());
-			endereco.setUf(responseViaCep.getUf());
-			endereco.setComplemento(clienteDTO.getComplemento());
-			endereco.setNumero(clienteDTO.getNumero());
-			cliente.setEndereco(endereco);
-			enderecoRepository.save(endereco);
-		} catch (HttpClientErrorException | HttpServerErrorException | UnknownHttpStatusCodeException e) {
-			throw new CEPNotFoundException(clienteDTO.getCep());
-		}
-
+		//Fim da verificação
+		
+		//Realiza a busca no ViaCep
+		Endereco enderecoViaCep = enderecoService.buscaCep(clienteDTO);
+		cliente.setEndereco(enderecoViaCep);
 		clienteRepository.save(cliente);
+		
 		ClienteDTO clienteSalvo = modelMapper.map(cliente, ClienteDTO.class);
 		clienteSalvo.setIdCliente(cliente.getIdCliente());
 		return clienteSalvo;
